pyproject.toml
[project]
name = "13-ai-agents-with-langgraph-and-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-core>=1.3.0",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
